+++ android/src/main/java/flutter/overlay/window/flutter_overlay_window/OverlayConstants.java
@@ -6,6 +6,6 @@ final public class OverlayConstants {
     static final String CHANNEL_TAG = "x-slayer/overlay_channel";
     static final String OVERLAY_TAG = "x-slayer/overlay";
     static final String MESSENGER_TAG = "x-slayer/overlay_messenger";
-    static final String CHANNEL_ID = "ForegroundServiceChannel";
+    static final String CHANNEL_ID = "Overlay Channel";
     static final int NOTIFICATION_ID = 4579;
 }